改变一些bug

+++ RoadToPeace/Assets/Source/Component/LifeComponent.cs
@@ -17,5 +17,5 @@ using Entitas.CodeGeneration.Attributes;
 //[Unique]
 public class LifeComponent : IComponent
 {
-    public int lifeValue;
+    public float lifeValue;
 }
